feat: solve day 4 part 1 in rust My result is 6189740 cards to be pulled in total. According to my benchmark, this takes about 30ms in my TypeScript program. Attempting this in Rust, reduced the amount of time to 18.72ms on average.
There was also a minor improvement by streaming the file and handling the data,
and not batching the entire file into memory first.
I'm not sure, if I'm happy with this result yet. But I don't see,
where I might have inefficiencies.

Not like 18.72ms are much for 6 million iterations, but still...^^
Anyways. This really shouldn't bother me, the task is solved. But I'm still wondering.